SentimentAnalyzer/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
decorator==4.4.2
Django==3.2.5
djangorestframework==3.12.4
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
joblib==1.0.1
moviepy==1.0.3
nltk==3.6.3
numpy==1.21.0
opencv-python==4.5.3.56
pandas==1.3.0
Pillow==8.3.1
proglog==0.1.9
pydub==0.25.1
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
threadpoolctl==2.2.0
tqdm==4.61.2
urllib3==1.26.6